LayoutTests/imported/w3c/web-platform-tests/css/css-values/random-in-animations.tentative-expected.txt


FAIL CSS Transitions: property <--number> from [100] to [random(300, 100)] at (-0.5) should be [0] assert_equals: expected "0 " but got "150 "
PASS CSS Transitions: property <--number> from [100] to [random(300, 100)] at (0) should be [100]
FAIL CSS Transitions: property <--number> from [100] to [random(300, 100)] at (0.5) should be [200] assert_equals: expected "200 " but got "50 "
FAIL CSS Transitions: property <--number> from [100] to [random(300, 100)] at (1) should be [300] assert_equals: expected "300 " but got "0 "
FAIL CSS Transitions: property <--number> from [100] to [random(300, 100)] at (1.5) should be [400] assert_equals: expected "400 " but got "- 50 "
FAIL CSS Animations: property <--number> from [100] to [random(300, 100)] at (-0.5) should be [0] assert_equals: expected "0 " but got "150 "
PASS CSS Animations: property <--number> from [100] to [random(300, 100)] at (0) should be [100]
FAIL CSS Animations: property <--number> from [100] to [random(300, 100)] at (0.5) should be [200] assert_equals: expected "200 " but got "50 "
FAIL CSS Animations: property <--number> from [100] to [random(300, 100)] at (1) should be [300] assert_equals: expected "300 " but got "0 "
FAIL CSS Animations: property <--number> from [100] to [random(300, 100)] at (1.5) should be [400] assert_equals: expected "400 " but got "- 50 "
FAIL Web Animations: property <--number> from [100] to [random(300, 100)] at (-0.5) should be [0] assert_equals: expected "0 " but got "150 "
PASS Web Animations: property <--number> from [100] to [random(300, 100)] at (0) should be [100]
FAIL Web Animations: property <--number> from [100] to [random(300, 100)] at (0.5) should be [200] assert_equals: expected "200 " but got "50 "
FAIL Web Animations: property <--number> from [100] to [random(300, 100)] at (1) should be [300] assert_equals: expected "300 " but got "0 "
FAIL Web Animations: property <--number> from [100] to [random(300, 100)] at (1.5) should be [400] assert_equals: expected "400 " but got "- 50 "
PASS CSS Transitions: property <font-weight> from [100] to [random(300, 100)] at (-0.5) should be [1]
PASS CSS Transitions: property <font-weight> from [100] to [random(300, 100)] at (0) should be [100]
PASS CSS Transitions: property <font-weight> from [100] to [random(300, 100)] at (0.5) should be [200]
PASS CSS Transitions: property <font-weight> from [100] to [random(300, 100)] at (1) should be [300]
PASS CSS Transitions: property <font-weight> from [100] to [random(300, 100)] at (1.5) should be [400]
PASS CSS Transitions with transition: all: property <font-weight> from [100] to [random(300, 100)] at (-0.5) should be [1]
PASS CSS Transitions with transition: all: property <font-weight> from [100] to [random(300, 100)] at (0) should be [100]
PASS CSS Transitions with transition: all: property <font-weight> from [100] to [random(300, 100)] at (0.5) should be [200]
PASS CSS Transitions with transition: all: property <font-weight> from [100] to [random(300, 100)] at (1) should be [300]
PASS CSS Transitions with transition: all: property <font-weight> from [100] to [random(300, 100)] at (1.5) should be [400]
PASS CSS Animations: property <font-weight> from [100] to [random(300, 100)] at (-0.5) should be [1]
PASS CSS Animations: property <font-weight> from [100] to [random(300, 100)] at (0) should be [100]
PASS CSS Animations: property <font-weight> from [100] to [random(300, 100)] at (0.5) should be [200]
PASS CSS Animations: property <font-weight> from [100] to [random(300, 100)] at (1) should be [300]
PASS CSS Animations: property <font-weight> from [100] to [random(300, 100)] at (1.5) should be [400]
PASS Web Animations: property <font-weight> from [100] to [random(300, 100)] at (-0.5) should be [1]
PASS Web Animations: property <font-weight> from [100] to [random(300, 100)] at (0) should be [100]
PASS Web Animations: property <font-weight> from [100] to [random(300, 100)] at (0.5) should be [200]
PASS Web Animations: property <font-weight> from [100] to [random(300, 100)] at (1) should be [300]
PASS Web Animations: property <font-weight> from [100] to [random(300, 100)] at (1.5) should be [400]
PASS Test custom property animation with random() endpoints
PASS Test font-weight animation with random() endpoints
